requirements.txt
alembic==1.9.4
asgiref==3.6.0
attrs==23.1.0
autopep8==2.0.1
azure-ai-formrecognizer==3.3.0
azure-ai-ml==1.9.0
azure-common==1.1.28
azure-core==1.29.2
azure-identity==1.14.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
azure-storage-blob==12.17.0
azure-storage-file-datalake==12.12.0
azure-storage-file-share==12.13.0
blinker==1.5
boto3==1.26.94
botocore==1.29.94
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
Django==4.1.7
djangorestframework==3.14.0
dnspython==2.4.2
et-xmlfile==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-MQTT==1.1.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
fpdf==1.7.2
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
image==1.5.33
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.2
marshmallow==3.20.1
msal==1.23.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.25.1
oauthlib==3.2.2
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
paho-mqtt==1.6.1
pdf2image==1.16.3
Pillow==9.4.0
portalocker==2.7.0
protobuf==4.24.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydash==5.1.2
PyJWT==2.8.0
pymongo==4.6.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-decouple==3.7
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.4
sqlparse==0.4.3
strictyaml==1.7.3
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.14
uWSGI==2.0.21
Werkzeug==2.2.3
WTForms==3.1.1
XlsxWriter==3.1.0
